refactor(pay-form): update component imports and test snap files
