+++ renderer/components/Pay/PayAmountFields.js
@@ -1,11 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { Box, Flex } from 'rebass'
-import { FormattedMessage, intlShape } from 'react-intl'
+import { FormattedMessage } from 'react-intl'
 import debounce from 'lodash/debounce'
 import { Keyframes } from 'react-spring/renderprops.cjs'
-import { Bar, TransactionFeeInput, Toggle, Label } from 'components/UI'
-import { CurrencyFieldGroup } from 'containers/UI'
+import { intlShape } from '@zap/i18n'
+import { TransactionFeeInput, Toggle, Label } from 'components/Form'
+import { Bar } from 'components/UI'
+import { CurrencyFieldGroup } from 'containers/Form'
 import { convert } from '@zap/utils/btc'
 import messages from './messages'
 import { PAY_FORM_STEPS } from './constants'
